requirements_ros.txt
absl-py==1.4.0
actionlib==1.14.0
angles==1.9.13
bondpy==1.8.6
cachetools==5.3.1
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
cmake==3.27.2
contourpy==1.1.0
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
Farama-Notifications==0.0.4
filelock==3.12.3
fonttools==4.42.1
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
gymnasium==0.29.1
idna==3.4
image-geometry==1.16.2
importlib-metadata==6.8.0
importlib-resources==6.0.1
interactive-markers==1.12.0
Jinja2==3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.5
laser_geometry==1.6.7
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
message-filters==1.16.0
moveit-commander==1.1.13
moveit-core==1.1.13
moveit-ros-planning-interface==1.1.13
moveit-ros-visualization==1.1.13
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2023.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.31.0
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.20
scipy==1.10.1
sensor-msgs==1.13.1
six==1.16.0
smclib==1.8.6
srdfdom==0.6.4
stable-baselines3==2.1.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
topic-tools==1.16.0
torch==2.0.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urdfdom-py==0.4.6
urllib3==1.26.16
Werkzeug==2.3.7
xacro==1.14.16
zipp==3.16.2
